fix: report an autostart uninstall that leaves a unit running, and document watch's exit status (#1101)
Partially closes #1080. **Does not close it** — see the scope note at
the end.

## What landed here

**`autostart uninstall` no longer reports success while a unit keeps
running.** Both modes discarded the systemctl result with `let _ =`, so
a unit that refused to stop was left running while uninstall printed its
removal lines and exited 0. The operator believes the stack is gone; it
comes back at the next boot.

Idempotence is kept, but by acting and checking the result rather than
probing a proxy first:

- Quadlet mode stops unconditionally. Measured: `stop` on a unit whose
fragment is on disk exits 0 in every state, and every unit in that loop
has its fragment on disk. An `is-active` probe is worse than redundant —
a service still coming up reports `activating`, which `is-active` calls
a failure, so a stack caught mid-start would be skipped and left running
while its units were deleted.
- Service mode disables whenever systemd knows the unit, keyed on
`is-active` exit **4** ("no such unit") — a number, not localised prose.
The unit *file* is not a usable proxy: measured, a fragment deleted out
of band leaves the unit loaded, enabled and running, and `disable --now`
is the only thing that clears it.

Verified end to end on real systemd: install, `rm` the fragment out of
band, `daemon-reload`, then `podup autostart uninstall` takes it from
`active` to `inactive`. `--purge` now runs before the failure surfaces,
since by then the unit files are gone and `installed_mode` would no
longer recognise the project.

**`stats` failures are visible.** They were logged at `debug!`,
invisible at the default `warn` floor. The exit code is unchanged.

**`watch`'s exit status is documented** as the one that means nothing —
a failed sync/rebuild/restart/exec is a warning and the session still
exits 0. That matches docker compose and is deliberate; what was missing
was saying so, in the exit-status table and on the library entry point.

Plus a mechanical split: the autostart tests moved to their own file
after the new cases pushed `mod.rs` one line over the 500-line hard
limit.

## What I backed out, and why

I tried twice to make `logs`/`stats`/`events` report a failure instead
of exiting 0, and the live lane rejected both.

The second rejection taught me I had misread the first. **The lane
prints every failing test, not only the unexpected ones.** Ten of the
fifteen it listed are already in `.github/podman-known-failures-5` — the
`run` family and the streaming reads nested-virt cannot deliver. Only
five were mine, and all five are `logs`.

So it was never the mid-stream error handling I narrowed away in the
second attempt. It was the two faults I had called unambiguous — a
replica lookup that fails, and a stream that will not open. Under
nested-virt on Podman 5.8.1 those happen transiently; turning them into
failures fails commands that worked. Real Podman 5.4.2 here is green on
the same commit, which is the whole reason the lane exists.

A third attempt without new information would be a guess. `logs` is back
to `develop`'s behaviour untouched, and the question moves to **#1104**,
which is specifically about podup not being able to tell a finished
libpod stream from a broken one.

## Scope: #1080 stays open

Of its ten points, five were already fixed by #1088 before I started — I
audited each against `develop` HEAD and posted the table on the issue.
Of the remaining five, this PR closes `autostart uninstall` and
documents `watch`. `logs`, `stats` and `events` exit codes are deferred
to #1104.

## Test plan

`cargo test --lib` 1260/1260, clippy `--all-targets --all-features`
clean, `RUSTDOCFLAGS="-D warnings" cargo doc` clean. Full integration
suite green locally against real Podman 5.4.2.

Four new autostart tests using a scriptable `SystemCtl` double (the
existing fake always returns 0, so it cannot express "not running" or
"the stop was refused"), including one asserting `is-active` is **not**
probed in the quadlet path so the bad gate cannot creep back. **All four
verified to fail without the fix** — the quadlet pair by stashing the
source, the service-mode pair by reverting just the `uninstall` body,
since those tests live inside `mod.rs`.

Part of #1080

---------
